fixed update , and ready for using proto to wal

+++ src/cli/commands.go
@@ -8,6 +8,7 @@ const (
 	ShowTables      Action = "show tables"
 	CreateTable     Action = "create table"
 	Insert          Action = "insert"
+	Update          Action = "update"
 	Delete          Action = "delete"
 	Get             Action = "get"
 	CommandNotFound Action = "not_found"
@@ -26,6 +27,10 @@ func parseCMD(cmd string) Action {
 
 	if strings.HasPrefix(cmd, "INSERT INTO") {
 		return Insert
+	}
+
+	if strings.HasPrefix(cmd, "UPDATE") {
+		return Update
 	}
 
 	if strings.HasPrefix(cmd, "DELETE FROM") {
